+++ src/main/java/com/springsecurity3withthymeleaf/asset/role/dao/RoleDao.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleDao extends JpaRepository< Role, Integer > {
+public interface RoleDao extends JpaRepository< Role, Long > {
     Role findByRoleName(String roleName);
 }
