+++ src/main/java/com/springsecurity3withthymeleaf/configuration/log_in_out_history/dao/FailureAttemptDao.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FailureAttemptDao extends JpaRepository< FailureAttempt, Integer > {
+public interface FailureAttemptDao extends JpaRepository< FailureAttempt, Long > {
 
   List< FailureAttempt > findByUsername(String username);
 }
