+++ src/main/java/com/springsecurity3withthymeleaf/configuration/log_in_out_history/service/impl/LogInOutHistoryServiceImpl.java
@@ -25,7 +25,7 @@ public class LogInOutHistoryServiceImpl implements LogInOutHistoryService {
   }
 
   @Cacheable
-  public LogInOutHistory findById(Integer id) {
+  public LogInOutHistory findById(Long id) {
     return logInOutHistoryDao.getReferenceById(id);
   }
